% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cid.R
\name{get_cid_line}
\alias{get_cid_line}
\title{Get Cycling Infrastructure data on lines}
\usage{
get_cid_line(type)
}
\arguments{
\item{type}{Type of infrastructure}
}
\value{
sf object
}
\description{
Get Cycling Infrastructure data on lines
}
